requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
rich==13.7.0
PyYAML==6.0.1